requirements.txt
jinja2
typeguard
spatialmath-python
matplotlib
mujoco>2.3.4
PyQt5